refactor(memory): move spreadActivation from the v2 engine into the substrate
